break code in modules and add more comments
