+++ hydra-redis/redis_usage
@@ -6,17 +6,17 @@ Requirements for redis_graph.py file according to my local env.
     pip install redisgraph       // you can change the client.py with your requirements
     and then first run the redis server and then run redis_graph.py file
     
+-----------------------------------------------------------------------------------------------------------------------    
     
-    
-After install redis-graph, redis and its python client 
+After install redis-graph, redis and its python client or you have already install them then
 you should do :
-    $cd redis
-    $src/redis-server --loadmodule /path_to_redis_graph/src/redisgraph.so
+    `$cd redis`
+    `$src/redis-server --loadmodule /path_to_redis_graph/src/redisgraph.so`
     like  this:
     src/redis-server --loadmodule /home/sandeep/demo/redis-module-graph/src/redisgraph.so
  it will run the redis server then you can run your redis file in different terminal but in same env.   
 
--------------------------------------------------------------------------------------------------------
+-----------------------------------------------------------------------------------------------------------------------
 
 There is also an alternate method for running the redis server 
 you should run `redis_setup.sh` after once read it.
